Allow param points to be floats

+++ js/canvas.js
@@ -200,8 +200,8 @@ function first_load() {
 
 	var origin = getParameterByName("origin"); // in format XxY
 	var terminus = getParameterByName("terminus"); // in format XxY
-	origin = {x:parseInt(origin.split("x")[0])+board.origin.x,y:parseInt(origin.split("x")[1])+board.origin.y};
-	terminus = {x:parseInt(terminus.split("x")[0])+board.origin.x,y:parseInt(terminus.split("x")[1])+board.origin.y};
+	origin = {x:parseFloat(origin.split("x")[0])+board.origin.x,y:parseFloat(origin.split("x")[1])+board.origin.y};
+	terminus = {x:parseFloat(terminus.split("x")[0])+board.origin.x,y:parseFloat(terminus.split("x")[1])+board.origin.y};
 	if(origin.x && origin.y){
 		console.log("Origin from param: ", origin);
 		template.setOrigin(origin);
